dashboard/requirements.txt
streamlit>=1.32.0
gspread>=6.0.0
google-auth>=2.28.0
plotly>=5.18.0
pandas>=2.2.0
